Word now holds long value instead of integer.

+++ src/main/java/com/github/justasbieliauskas/rmvm/cpu/CPU.java
@@ -15,5 +15,5 @@ public interface CPU
      *
      * @return map representation
      */
-    Map<String, Integer> toMap();
+    Map<String, Long> toMap();
 }
